# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Todo Microservice API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /calendar/sources/import:
    post:
      security:
        - bearerAuth: []
      summary: Import a calendar source from an uploaded ICS file
      operationId: importCalendarSource
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - category
              properties:
                file:
                  type: string
                  format: binary
                category:
                  type: string
                display_name:
                  type: string
      responses:
        "201":
          description: Calendar source imported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarImportResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
  /calendar/sources:
    get:
      security:
        - bearerAuth: []
      summary: Get calendar sources for the current user
      operationId: getCalendarSources
      responses:
        "200":
          description: Calendar sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarSource"
        "401":
          description: Unauthorized
  /calendar/sources/link:
    post:
      security:
        - bearerAuth: []
      summary: Add a linked ICS calendar source
      operationId: createLinkedCalendarSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewCalendarLinkSource"
      responses:
        "201":
          description: Linked calendar source imported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarImportResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
  /calendar/sources/{sourceId}:
    get:
      security:
        - bearerAuth: []
      summary: Get a calendar source by ID
      operationId: getCalendarSourceById
      parameters:
        - in: path
          name: sourceId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Calendar source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarSource"
        "404":
          description: Calendar source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      security:
        - bearerAuth: []
      summary: Rename a calendar source
      operationId: updateCalendarSource
      parameters:
        - in: path
          name: sourceId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCalendarSource"
      responses:
        "200":
          description: Updated calendar source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarSource"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Calendar source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      security:
        - bearerAuth: []
      summary: Delete a calendar source and its imported events
      operationId: deleteCalendarSource
      parameters:
        - in: path
          name: sourceId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Calendar source deleted successfully
        "404":
          description: Calendar source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /calendar/sources/{sourceId}/refresh:
    post:
      security:
        - bearerAuth: []
      summary: Refresh a linked calendar source
      operationId: refreshCalendarSource
      parameters:
        - in: path
          name: sourceId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Calendar source refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarImportResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Calendar source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /calendar/events:
    get:
      security:
        - bearerAuth: []
      summary: Get imported calendar events for the current user
      operationId: getCalendarEvents
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: sourceId
          required: false
          schema:
            type: string
            format: uuid
        - in: query
          name: cursor
          required: false
          description: Anchor timestamp for cursor-based schedule pagination.
          schema:
            type: string
            format: date-time
        - in: query
          name: direction
          required: false
          description: Fetch events before or after the cursor.
          schema:
            type: string
            enum: [before, after]
        - in: query
          name: limit
          required: false
          description: Maximum number of events to return for cursor-based queries.
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
        "401":
          description: Unauthorized
        "404":
          description: Calendar source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /calendar/events/{eventId}:
    get:
      security:
        - bearerAuth: []
      summary: Get a calendar event by ID
      operationId: getCalendarEventById
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Calendar event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
        "404":
          description: Calendar event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /calendar/upcoming:
    get:
      security:
        - bearerAuth: []
      summary: Get upcoming imported calendar events for the current user
      operationId: getUpcomingCalendarEvents
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Upcoming calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Something went wrong"
...
    CalendarSource:
      type: object
      required:
        - id
        - user_id
        - kind
        - display_name
        - category
        - import_mode
        - refresh_state
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        kind:
          type: string
          example: ics_file
        display_name:
          type: string
          example: Work Calendar
        category:
          type: string
          example: Work
        import_mode:
          type: string
          example: upload
        source_url:
          type: string
          nullable: true
          example: https://calendar.example.com/feed.ics
        refresh_state:
          type: string
          example: imported
        last_synced_at:
          type: string
          format: date-time
        last_refresh_attempt_at:
          type: string
          format: date-time
        last_refresh_error:
          type: string
          nullable: true
        etag:
          type: string
          nullable: true
        last_modified:
          type: string
          format: date-time
        next_refresh_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NewCalendarLinkSource:
      type: object
      required:
        - url
        - category
      properties:
        url:
          type: string
          format: uri
        category:
          type: string
        display_name:
          type: string
    UpdateCalendarSource:
      type: object
      required:
        - display_name
        - category
      properties:
        category:
          type: string
        display_name:
          type: string
    CalendarEvent:
      type: object
      required:
        - id
        - source_id
        - external_uid
        - title
        - description
        - location
        - starts_at
        - ends_at
        - all_day
        - status
        - timezone
        - source_display_name
      properties:
        id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        external_uid:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        all_day:
          type: boolean
        status:
          type: string
        timezone:
          type: string
        recurrence_summary:
          type: string
          nullable: true
        source_display_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CalendarImportResponse:
      type: object
      required:
        - source
        - imported_event_count
      properties:
        source:
          $ref: "#/components/schemas/CalendarSource"
        imported_event_count:
          type: integer
          format: int32
...